feat/refactor: refactor imports and add correct forgot password implementation
