+++ frontend/src/app/app.scss
@@ -68,3 +68,10 @@
     }
   }
 }
+
+.guest-content{
+  width: 100%;
+  height: 100%;
+  display: flex;
+
+}
