+++ frontend/src/app/modules/auth/auth.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { Login } from './components/login/login';
 import { Register } from './components/register/register';
-import { PasswordReset } from './password-reset/password-reset';
+import { PasswordReset } from './components/password-reset/password-reset';
 
 export const AUTH_ROUTES: Routes = [
   { path: 'login', component: Login },
