+++ frontend/src/app/modules/auth/components/password-reset/password-reset.html
@@ -81,8 +81,8 @@
   (visibleChange)="visible.set($event)"
   type="success"
   [title]="translationService.translate('passwordReset.dialog.successTitle')"
-  [messages]="[translationService.translate('passwordReset.dialog.sent')]"
-  [buttonLabel]="translationService.translate('passwordReset.dialog.signIn')"
+  [messages]="passwordSended() ? [translationService.translate('passwordReset.dialog.success')] : [translationService.translate('passwordReset.dialog.sent')]"
+  [buttonLabel]="passwordSended() ? translationService.translate('passwordReset.dialog.signIn') : translationService.translate('passwordReset.dialog.checkemail')"
   buttonIcon="pi pi-arrow-right"
   buttonIconPos="right"
   (buttonClick)="redirectToLoginPage()"
