+++ frontend/src/app/modules/guest/components/landing-page-component/landing-page-component.scss
@@ -32,9 +32,9 @@
   margin-top: 2rem;
 }
 :host {
+  width: 100%;
+  height: 100%;
   display: flex;
-  justify-content: center;
-
   flex-direction: column;
 }
 p-card {
